Move module online info methods into separate behavior
